Initial commit to support Casual Clusters
This is little flaky, but I have made successful writes in testing. The flakiness seems to be that I'm not always getting a writable leader. The TTLs are very short on my test database so I'm thinking there might be a race between fetching the routing table and sending the queries. So far this is somewhat magical but does not break the existing API. The implementation parses the statements to see if a write is needed and will get one. Otherwise it uses the default connection as configured in the ClientBuilder.