Make `mixed` for CypherList/CypherMap explicit
Psalm automatically fills in `<mixed>` for the missing type parameter.
However a lot of PHP projects use PHPStan which has no way of doing
this. This makes it impossible to reuse the `OGMTypes` type and as a
result makes it impossible to annotate types that match what Neo4j is
providing. Being explicit about `mixed` makes the behaviour between
Psalm and PHPStan the same.